patterns/bootstrap-standards/after/requirements.txt
pandas==2.1.4
pyyaml==6.0.1
pytest==7.4.4
